sqlite class DBHelper

+++ app/src/main/java/ru/serv_techno/delivery_st/Order.java
@@ -6,16 +6,15 @@ package ru.serv_techno.delivery_st;
 public class Order {
 
     int id;
-    int price;
+    float price;
     int created_at;
     int number_person;
     int delivery;
     String orderClientName;
     String orderClientPhone;
     String orderClientAddress;
-    //orderProducts
 
-    Order(int _id, int _price, int _created_at, int _number_person, int _delivery, String _orderClientName, String _orderClientPhone, String _orderClientAddress){
+    Order(int _id, float _price, int _created_at, int _number_person, int _delivery, String _orderClientName, String _orderClientPhone, String _orderClientAddress){
         id = _id;
         price = _price;
         created_at = _created_at;
